requirements.txt
rich
pyyaml
pymongo
discord
cogwatch
colorlog
requests
python-dotenv